Move LanguageResolver into own namespace and fix phpdoc to use FQCN
