+++ eZ/Bundle/EzPublishCoreBundle/Resources/config/helpers.yml
@@ -60,5 +60,5 @@ services:
             - "@ezpublish.api.repository"
 
     ezpublish.helper.language_resolver:
-        class: eZ\Publish\Core\Repository\Helper\LanguageResolver
+        class: eZ\Publish\Core\Repository\SiteAccessAware\Language\LanguageResolver
         arguments: ['$languages$']
